+++ src/packages/string_more.rs
@@ -116,9 +116,7 @@ mod string_functions {
     #[rhai_fn(name = "to_upper")]
     pub fn to_upper_char(character: char) -> char {
         let mut stream = character.to_uppercase();
-        let ch = stream
-            .next()
-            .expect("never fails because there should be at least one character");
+        let ch = stream.next().expect("at least one character");
         if stream.next().is_some() {
             character
         } else {
@@ -134,7 +132,7 @@ mod string_functions {
         let mut stream = character.to_lowercase();
         let ch = stream
             .next()
-            .expect("never fails because there should be at least one character");
+            .expect("there should be at least one character");
         if stream.next().is_some() {
             character
         } else {