+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Enhancements
 * `Scope::set_value` now takes anything that implements `Into<Cow<str>>`.
 * Added `Scope::is_constant` to check if a variable is constant.
 * Added `Scope::set_or_push` to add a new variable only if one doesn't already exist.
+* `Engine::register_type_XXX` are now available even under `no_object`.
 
 
 Version 1.0.2